refactor(botonic-react): create Textarea component for user input
